Move code into domain/resourcecontainer/project.
