+++ src/main/kotlin/org/wycliffeassociates/otter/common/domain/resourcecontainer/project/IProjectReader.kt
@@ -1,9 +1,10 @@
-package org.wycliffeassociates.otter.common.domain.resourcecontainer
+package org.wycliffeassociates.otter.common.domain.resourcecontainer.project
 
 import org.wycliffeassociates.otter.common.collections.tree.Tree
+import org.wycliffeassociates.otter.common.domain.resourcecontainer.ImportResourceContainer
 import org.wycliffeassociates.resourcecontainer.ResourceContainer
 import org.wycliffeassociates.resourcecontainer.entity.Project
 
-interface IProjectFileReader {
+interface IProjectReader {
     fun constructProjectTree(container: ResourceContainer, project: Project): Pair<ImportResourceContainer.Result, Tree>
 }
