+++ src/main/kotlin/org/wycliffeassociates/otter/common/domain/resourcecontainer/project/markdown/ParseMd.kt
@@ -1,4 +1,4 @@
-package org.wycliffeassociates.otter.common.domain.md
+package org.wycliffeassociates.otter.common.domain.resourcecontainer.project.markdown
 
 import java.io.BufferedReader
